Broke apart the tasks bootstrap module to be one module per task.
Fixed typing errors.

Added the ability to skip tests.

Added the ability to verify that tasks set all expected outputs.
