+++ .vscode/settings.json
@@ -19,12 +19,15 @@
     "dataclass",
     "dataclasses",
     "dearpygui",
+    "deepskyblue",
     "DLFL",
     "drawlist",
     "dunder",
     "Eberly",
     "Fira",
     "fmap",
+    "fontcolor",
+    "fontname",
     "FPTTL",
     "Framebuffer",
     "Fullscreen",
@@ -61,6 +64,7 @@
     "perp",
     "pfaults",
     "PICOMETER",
+    "plantext",
     "polyline",
     "prec",
     "psutil",
