+++ agents_playground/core/privileged.py
@@ -7,13 +7,18 @@ import os
 from typing import Any, Callable
 
 
+def running_as_root() -> bool:
+    """Determine if the process is running as root."""
+    return os.geteuid() == 0
+
+
 def require_root(func) -> Callable:
     """Only call the decorated function if the effective user is root."""
 
     @wraps(func)
     def wrapper(*args, **kargs) -> Any:
         result = None
-        if os.geteuid() == 0:
+        if running_as_root():
             result = func(*args, **kargs)
         return result
