+++ agents_playground/gpu/renderers/agent_renderer.py
@@ -27,8 +27,7 @@ from agents_playground.gpu.shader_configuration.default_shader_configuration_bui
 )
 from agents_playground.gpu.shaders import load_shader
 from agents_playground.scene import Scene
-from agents_playground.simulation.context import SimulationContextBuilder
-from agents_playground.spatial.matrix.matrix import Matrix, MatrixOrder
+from agents_playground.simulation.simulation_context_builder import SimulationContextBuilder
 from agents_playground.spatial.mesh import MeshBuffer, MeshData
 from agents_playground.spatial.transformation.configuration import TransformationConfiguration
