+++ agents_playground/simulation/simulation_context_builder.py
@@ -44,9 +44,17 @@ class SimulationContextBuilder:
     def scene(self) -> Scene:
         return self._scene.unwrap_or_throw(self._error_msg("scene"))
 
+    @scene.setter
+    def scene(self, other: Scene) -> None:
+        self._scene = Something(other)
+
     @property
     def canvas(self) -> WgpuWidget:
         return self._canvas.unwrap_or_throw(self._error_msg("canvas"))
+
+    @canvas.setter
+    def canvas(self, other: WgpuWidget) -> None:
+        self._canvas = Something(other)
 
     @property
     def render_texture_format(self) -> str:
@@ -54,53 +62,45 @@ class SimulationContextBuilder:
             self._error_msg("render_texture_format")
         )
 
-    @property
-    def device(self) -> GPUDevice:
-        return self._device.unwrap_or_throw(self._error_msg("device"))
-
-    @property
-    def renderers(self) -> dict[str, GPURenderer]:
-        return self._renderers.unwrap_or_throw(self._error_msg("renderers"))
-
-    @property
-    def mesh_registry(self) -> MeshRegistry:
-        return self._mesh_registry.unwrap_or_throw(self._error_msg("mesh_registry"))
-
-    @property
-    def extensions(self) -> dict[Any, Any]:
-        return self._extensions.unwrap_or_throw(self._error_msg("extensions"))
-
-    @property
-    def uniforms(self) -> UniformRegistry:
-        return self._uniforms.unwrap_or_throw(self._error_msg("uniforms"))
-
-    @scene.setter
-    def scene(self, other: Scene) -> None:
-        self._scene = Something(other)
-
-    @canvas.setter
-    def canvas(self, other: WgpuWidget) -> None:
-        self._canvas = Something(other)
-
     @render_texture_format.setter
     def render_texture_format(self, other: str) -> None:
         self._render_texture_format = Something(other)
+
+    @property
+    def device(self) -> GPUDevice:
+        return self._device.unwrap_or_throw(self._error_msg("device"))
 
     @device.setter
     def device(self, other: GPUDevice) -> None:
         self._device = Something(other)
 
+    @property
+    def renderers(self) -> dict[str, GPURenderer]:
+        return self._renderers.unwrap_or_throw(self._error_msg("renderers"))
+
     @renderers.setter
     def renderers(self, other: dict[str, GPURenderer]) -> None:
         self._renderers = Something(other)
+
+    @property
+    def mesh_registry(self) -> MeshRegistry:
+        return self._mesh_registry.unwrap_or_throw(self._error_msg("mesh_registry"))
 
     @mesh_registry.setter
     def mesh_registry(self, other: MeshRegistry) -> None:
         self._mesh_registry = Something(other)
 
+    @property
+    def extensions(self) -> dict[Any, Any]:
+        return self._extensions.unwrap_or_throw(self._error_msg("extensions"))
+
     @extensions.setter
     def extensions(self, other: dict[Any, Any]) -> None:
         self._extensions = Something(other)
+
+    @property
+    def uniforms(self) -> UniformRegistry:
+        return self._uniforms.unwrap_or_throw(self._error_msg("uniforms"))
 
     @uniforms.setter
     def uniforms(self, other: UniformRegistry) -> None:
@@ -112,7 +112,7 @@ class SimulationContextBuilder:
         return all(items_set)
 
     def create_context(self) -> SimulationContext:
-        return SimulationContext(
+        sc: SimulationContext = SimulationContext(
             scene=self.scene,
             canvas=self.canvas,
             render_texture_format=self.render_texture_format,
@@ -121,3 +121,4 @@ class SimulationContextBuilder:
             extensions=self.extensions,
             uniforms=self.uniforms,
         )
+        return sc
